# CORE-207: optimize stringToTypedAttribute in TSV upload