This PR is an optimization found while researching CORE-207. This should be performance only; it has no functional impact to end users.

The `stringToTypedAttribute()` function is called for each cell in an uploaded TSV; this is a lot of invocations so any optimization here is useful.

1. *check the cell's first character*. Previously, we attempted to parse the cell as a number, a boolean, and a reference before giving up and calling it a string. In this PR, we grab the first non-whitespace character of the cell and compare it to a known list of possible first characters for the non-string datatypes - for instance, a boolean will always start with "t" or "f". This is faster than trying to parse the whole cell and results in a ~2.4x speedup in our new benchmark.

Raw benchmark numbers:

  | ops/sec | error +/- | vs.baseline | avg ops/sec | avg vs. baseline
-- | -- | -- | -- | -- | --
baseline (648c5f274c4af9541f1d71c4c50db6961aecc3ed) | 13,426.94 | 598.82 |   | 13,227.86 |  
  | 12,690.29 | 564.33 | 94.51% |   |  
  | 13,566.35 | 277.55 | 101.04% |   |  
first-character checks (ed275ec29fd487778f035b9a515377528ee00150) | 32,275.95 | 660.82 | 240.38% | 32,325.38 | 244.37%
  | 32,994.04 | 912.17 | 245.73% |   |  
  | 31,706.16 | 1,068.82 | 236.14% |   |  
more Scala-y (6ad61e4898a3b1cbc3cc91d210ce029b47196b05) | 32,781.50 | 564.73 | 244.15% | 32,126.14 | 242.87%
  | 31,982.75 | 1,143.94 | 238.20% |   |  
  | 31,614.16 | 634.61 | 235.45% |   |  

---

Have you read [CONTRIBUTING.md](../CONTRIBUTING.md) lately? If not, do that first.

I, the developer opening this PR, do solemnly pinky swear that:

- [ ] I've followed [the instructions](https://github.com/broadinstitute/firecloud-orchestration/blob/develop/CONTRIBUTING.md#api-changes) if I've made any changes to the API, _especially_ if they're breaking changes
- [ ] I've updated the [FISMA documentation](https://github.com/broadinstitute/firecloud-orchestration/blob/develop/CONTRIBUTING.md#fisma-documentation-changes) if I've made any security-related changes, including auth, encryption, or auditing

In all cases:

- [ ] Get two thumbsworth of review and PO signoff if necessary
- [ ] Verify all tests go green
- [ ] Squash and merge. Make sure your branch deletes; GitHub should do this for you.
- [ ] Test this change deployed correctly and works on dev environment after deployment